requirements.txt
pandas
numpy
plotly
dash
pyserial
firebase-admin
python-dotenv